Remove getAST non-static member on RecordTypeFunction.  Don't want to risk breaking any existing code so isolate this functionality into the new code paths represented by the static members

+++ ads-ql/RecordType.hh
@@ -173,7 +173,7 @@ public:
   {
     return *(double *) (buffer.Ptr + mOffset);
   }
-  int32_t getArrayDouble(RecordBuffer buffer, int idx) const
+  double getArrayDouble(RecordBuffer buffer, int idx) const
   {
     return ((double *) (buffer.Ptr + mOffset))[idx];
   }
